+++ base/options.jl
@@ -73,6 +73,8 @@ struct JLOptions
     target_sanitize_memory::Int8
     target_sanitize_thread::Int8
     target_sanitize_address::Int8
+    sysimage_prelink::Int8
+    output_prelinked::Ptr{UInt8}
 end
 
 # This runs early in the sysimage when `!=` is not defined yet