+++ src/jloptions.h
@@ -77,6 +77,8 @@ typedef struct {
     int8_t target_sanitize_memory;
     int8_t target_sanitize_thread;
     int8_t target_sanitize_address;
+    int8_t sysimage_prelink;
+    const char *output_prelinked;
 } jl_options_t;
 
 #endif